OpenHuntingData Processing Tool MVP
A tool has been written in Go to download and process an example data
source, particularly sources/US/MT/turkey.json. Instructions for
building and running the tool can be found in the README.md file. It is
only a minimal implementation that can process a single data file, but
has much recyclable code that can be further genericized.